Change CMSIS from rev. 5.1.1 to rev. 6.1.0

+++ Library/StdDriver/src/utcpd.c
@@ -53,7 +53,7 @@ extern "C"
   */
 int32_t UTCPD_GetAlertStatus(int port, int* i32AlertSts)
 {
-    return tcpc_addr_read16(port, NULL, TCPC_REG_ALERT, i32AlertSts);	
+    return tcpc_addr_read16(port, (int) NULL, TCPC_REG_ALERT, i32AlertSts);	
 }
 
 /**
@@ -383,7 +383,7 @@ int32_t UTCPD_DisableFaultStatusMask(int port, int mask_clr)
   */
 int32_t UTCPD_GetFaultStatus(int port, int* pi32RegData)
 {
-	  return tcpc_addr_read16(port, NULL, TCPC_REG_FAULT_STATUS, pi32RegData);
+	  return tcpc_addr_read16(port, (int) NULL, TCPC_REG_FAULT_STATUS, pi32RegData);
 }
 
 /**
@@ -431,7 +431,7 @@ int32_t UTCPD_ClearFaultStatus(int port, int FaultStClr)
 int32_t UTCPD_SetRoleCtrl(int port, uint32_t u32DrpToggle, uint32_t u32Rpvalue, uint32_t u32CC2, uint32_t u32CC1)
 {
 		uint8_t reg_set = u32DrpToggle | u32Rpvalue | u32CC2 | u32CC1; 
-		return tcpc_addr_write16(port, NULL, TCPC_REG_ROLE_CTRL, reg_set);	 
+		return tcpc_addr_write16(port, (int) NULL, TCPC_REG_ROLE_CTRL, reg_set);	 
 }
 
 /**
@@ -464,7 +464,7 @@ int32_t UTCPD_GetRoleCtrl(int port, uint32_t* pu32DrpToggle, uint32_t* pu32CC1, 
 	int i32Reg;
 	int32_t rv;
 
-	rv = tcpc_addr_read32(port, NULL, TCPC_REG_ROLE_CTRL, &i32Reg);
+	rv = tcpc_addr_read32(port, (int) NULL, TCPC_REG_ROLE_CTRL, &i32Reg);
 	*pu32DrpToggle = i32Reg & UTCPD_ROLCTL_DRP_Msk;
 	*pu32Rpvalue = i32Reg & UTCPD_ROLCTL_RPVALUE_Msk;
 	*pu32CC2 = i32Reg & UTCPD_ROLCTL_CC2_Msk;
@@ -489,7 +489,7 @@ int32_t UTCPD_GetRoleCtrl(int port, uint32_t* pu32DrpToggle, uint32_t* pu32CC1, 
 int32_t UTCPD_SetTypeCPortCtrl(int port, uint32_t u32BistMode, uint32_t u32Orient)
 {
 	uint8_t reg_set = u32BistMode | u32Orient; 
-	return tcpc_addr_write16(port, NULL, TCPC_REG_TCPC_CTRL, reg_set);	
+	return tcpc_addr_write16(port, (int) NULL, TCPC_REG_TCPC_CTRL, reg_set);	
 }
 
 
@@ -510,7 +510,7 @@ int32_t UTCPD_GetTypeCPortCtrl(int port, uint32_t* pu32BistMode, uint32_t* pu32O
 	int i32Reg;
 	int32_t rv;
 	
-	rv = tcpc_addr_read32(port, NULL, TCPC_REG_TCPC_CTRL, &i32Reg);
+	rv = tcpc_addr_read32(port, (int) NULL, TCPC_REG_TCPC_CTRL, &i32Reg);
 	*pu32BistMode = i32Reg & UTCPD_TCPCCTL_BISTEN;
 	*pu32Orient = i32Reg & UTCPD_TCPCCTL_ORIENT;
 	return rv;	
@@ -535,7 +535,7 @@ int32_t UTCPD_GetTypeCPortCtrl(int port, uint32_t* pu32BistMode, uint32_t* pu32O
 int32_t UTCPD_IsssueCmd(int port, uint32_t cmd)
 {
 
-	return tcpc_addr_write16(port, NULL, TCPC_REG_COMMAND, cmd);	
+	return tcpc_addr_write16(port, (int) NULL, TCPC_REG_COMMAND, cmd);	
 }
 
 
@@ -569,7 +569,7 @@ int32_t UTCPD_GetCCSts(int port, uint32_t* pu32Look4Con, uint32_t* pu32ConRlt, u
 	int i32Reg;
 	int32_t rv;
 
-	rv = tcpc_addr_read32(port, NULL, TCPC_REG_CC_STATUS, &i32Reg);
+	rv = tcpc_addr_read32(port, (int) NULL, TCPC_REG_CC_STATUS, &i32Reg);
 	*pu32Look4Con = (i32Reg & TCPC_REG_CC_STATUS_LOOK4CONNECTION_MASK);
 	*pu32ConRlt = (i32Reg & TCPC_REG_CC_STATUS_CONNECT_RESULT_MASK);
 	*pu32CC2Sts = (i32Reg & TCPC_REG_CC_STATUS_CC2_STATE_MASK);
@@ -600,7 +600,7 @@ int32_t UTCPD_GetPwrSts(int port, uint32_t* pu32VBUSDetEn, uint32_t* pu32VBUSPre
 	int i32Reg;
 	int32_t rv;
 
-	rv = tcpc_addr_read32(port, NULL, TCPC_REG_POWER_STATUS, &i32Reg);
+	rv = tcpc_addr_read32(port, (int) NULL, TCPC_REG_POWER_STATUS, &i32Reg);
     *pu32VBUSDetEn = i32Reg & UTCPD_PWRSTS_VBPSDTEN_Msk;
     *pu32VBUSPresent = i32Reg & UTCPD_PWRSTS_VBPS_Msk;
     *pu32VCONNPresent = i32Reg & UTCPD_PWRSTS_VCPS_Msk;
@@ -630,7 +630,7 @@ int32_t UTCPD_GetPwrStsExt(int port, uint32_t* pu32DbgAccessory, uint32_t* pu32S
 	int i32Reg;
 	int32_t rv;
 
-	rv = tcpc_addr_read32(port, NULL, TCPC_REG_POWER_STATUS, &i32Reg);
+	rv = tcpc_addr_read32(port, (int) NULL, TCPC_REG_POWER_STATUS, &i32Reg);
 	*pu32DbgAccessory = i32Reg & UTCPD_PWRSTS_DACON_Msk;
 	*pu32SrcNonDefVBUS = i32Reg & UTCPD_PWRSTS_SRHV_Msk;
 	*pu32SrcDefVBUS = i32Reg & UTCPD_PWRSTS_SRVB_Msk;
@@ -659,7 +659,7 @@ int32_t UTCPD_GetFaultSts(int port, uint32_t* pu32VBUSOverCurr, uint32_t* pu32VB
 	int i32Reg;
 	int32_t rv;
 	
-	rv = tcpc_addr_read32(port, NULL, TCPC_REG_FAULT_STATUS, &i32Reg);
+	rv = tcpc_addr_read32(port, (int) NULL, TCPC_REG_FAULT_STATUS, &i32Reg);
 	*pu32VBUSOverCurr = i32Reg & UTCPD_FUTSTS_VBOCFUT_Msk;
 	*pu32VBUSOverVolt = i32Reg & UTCPD_FUTSTS_VBOVFUT_Msk;
 	*pu32VCONNOverCurr = i32Reg & UTCPD_FUTSTS_VCOCFUT_Msk;
@@ -690,7 +690,7 @@ int32_t UTCPD_GetFaultStsExt(int port, uint32_t* pu32ForceOffFat, uint32_t* pu32
 	int i32Reg;
 	int32_t rv;
 	
-	rv = tcpc_addr_read32(port, NULL, TCPC_REG_FAULT_STATUS, &i32Reg);
+	rv = tcpc_addr_read32(port, (int) NULL, TCPC_REG_FAULT_STATUS, &i32Reg);
 	*pu32ForceOffFat = i32Reg & UTCPD_FUTSTS_FOFFVB_Msk;
     *pu32AutoDiscFat = i32Reg & UTCPD_FUTSTS_ADGFAL_Msk;
 	*pu32ForceDiscFat = i32Reg & UTCPD_FUTSTS_FDGFAL_Msk;
@@ -721,7 +721,7 @@ int32_t UTCPD_GetFaultStsExt(int port, uint32_t* pu32ForceOffFat, uint32_t* pu32
 int32_t UTCPD_SetMsgHeaderInfo(int port, uint32_t u32DataRole, uint32_t u32Revision, uint32_t u32PwrRole)
 {
 	uint8_t reg_set = u32DataRole | u32Revision | u32PwrRole; 
-	return tcpc_addr_write16(port, NULL, TCPC_REG_MSG_HDR_INFO, reg_set);	
+	return tcpc_addr_write16(port, (int) NULL, TCPC_REG_MSG_HDR_INFO, reg_set);	
 }
 
 /**
@@ -745,7 +745,7 @@ int32_t UTCPD_SetMsgHeaderInfo(int port, uint32_t u32DataRole, uint32_t u32Revis
 int32_t UTCPD_SetRecDetect(int port, uint32_t u32RegData)
 {
 	uint8_t reg_set = u32RegData;
-	return tcpc_addr_write16(port, NULL, TCPC_REG_RX_DETECT, reg_set);	
+	return tcpc_addr_write16(port, (int) NULL, TCPC_REG_RX_DETECT, reg_set);	
 }
 
 /**
@@ -782,7 +782,7 @@ int32_t UTCPD_SetVBUSAlarm(int port, uint32_t u32AlarmH, uint32_t u32AlarmL)
   */
 int32_t UTCPD_SetSnkDisconnect(int port, uint32_t u32SnkDiscVolt)
 {
-	return tcpc_addr_write16(port, NULL, TCPC_REG_VBUS_SINK_DISCONNECT_THRESH, u32SnkDiscVolt);	
+	return tcpc_addr_write16(port, (int) NULL, TCPC_REG_VBUS_SINK_DISCONNECT_THRESH, u32SnkDiscVolt);	
 }
 
 /**
@@ -798,7 +798,7 @@ int32_t UTCPD_SetSnkDisconnect(int port, uint32_t u32SnkDiscVolt)
   */
 int32_t UTCPD_SetStopDischargeVolt(int port, uint32_t u32StopDischgVolt)
 {
-  return tcpc_addr_write16(port, NULL, TCPC_REG_VBUS_STOP_DISCHARGE_THRESH, u32StopDischgVolt);	
+  return tcpc_addr_write16(port, (int) NULL, TCPC_REG_VBUS_STOP_DISCHARGE_THRESH, u32StopDischgVolt);	
 }
 
 
@@ -929,9 +929,9 @@ void UTCPD_vconn_enable_oc_fault(int port)
 void UTCPD_vconn_mux_selection(int port, uint32_t cc1vcensel, uint32_t cc2vcensel)
 {
 	int data; 
-	tcpc_addr_read16(port, NULL, UTCPD_MUXSEL, &data);
+	tcpc_addr_read16(port, (int) NULL, UTCPD_MUXSEL, &data);
 	data = (data & ~(CC2VCENS | CC1VCENS)) | ((cc1vcensel<<24) | (cc2vcensel<<28)); 
-    tcpc_addr_write32(port, NULL, UTCPD_MUXSEL, (unsigned int)data);
+    tcpc_addr_write32(port, (int) NULL, UTCPD_MUXSEL, (unsigned int)data);
 }
 
 /**
@@ -953,9 +953,9 @@ void UTCPD_vconn_mux_selection(int port, uint32_t cc1vcensel, uint32_t cc2vcense
 void UTCPD_vconn_configure_oc_detection_soruce(int port, uint32_t u32Src)
 {
 	int data; 
-	tcpc_addr_read16(port, NULL, UTCPD_MUXSEL, &data);
+	tcpc_addr_read16(port, (int) NULL, UTCPD_MUXSEL, &data);
 	data = (data & ~VCOCS) | (u32Src<<4); 
-    tcpc_addr_write16(port, NULL, UTCPD_MUXSEL, data);
+    tcpc_addr_write16(port, (int) NULL, UTCPD_MUXSEL, data);
 }
 
 /* ============  VBUS SRCEN Polarity ==========*/
@@ -1014,9 +1014,9 @@ void UTCPD_vbus_discharge_polarity_active_high(int port)
 void UTCPD_vbus_configure_oc_soruce(int port, uint32_t u32Src)
 {
 	int data; 
-	tcpc_addr_read16(port, NULL, UTCPD_MUXSEL, &data);
+	tcpc_addr_read16(port, (int) NULL, UTCPD_MUXSEL, &data);
 	data = (data & ~VBOCS) | (u32Src<<0); 
-    tcpc_addr_write16(port, NULL, UTCPD_MUXSEL, data);
+    tcpc_addr_write16(port, (int) NULL, UTCPD_MUXSEL, data);
 }
 
 /* VBUS ovp fault */
@@ -1043,7 +1043,7 @@ void UTCPD_vbus_enable_forceoff_fault(int port)
 uint32_t UTCPD_vbus_is_source(int port)
 {
     int data; 
-    tcpc_addr_read16(port, NULL, TCPC_REG_POWER_STATUS, &data);
+    tcpc_addr_read16(port, (int) NULL, TCPC_REG_POWER_STATUS, &data);
 	if (data & TCPC_REG_POWER_STATUS_SOURCING_VBUS)
         return 1; 
 	else
@@ -1053,7 +1053,7 @@ uint32_t UTCPD_vbus_is_source(int port)
 uint32_t UTCPD_vbus_is_sink(int port)
 {
     int data; 
-    tcpc_addr_read16(port, NULL, TCPC_REG_POWER_STATUS, &data);
+    tcpc_addr_read16(port, (int) NULL, TCPC_REG_POWER_STATUS, &data);
 	if (data & TCPC_REG_POWER_STATUS_SINKING_VBUS)
         return 1; 
 	else
@@ -1063,7 +1063,7 @@ uint32_t UTCPD_vbus_is_sink(int port)
 uint32_t UTCPD_vbus_is_source_hv(int port)
 {
     int data; 
-    tcpc_addr_read16(port, NULL, TCPC_REG_POWER_STATUS, &data);
+    tcpc_addr_read16(port, (int) NULL, TCPC_REG_POWER_STATUS, &data);
 	if (data & TCPC_REG_POWER_STATUS_SOURCING_HIGH_VBUS)
         return 1; 
 	else
@@ -1103,9 +1103,9 @@ void UTCPD_frs_tx_polarity_active_high(int port)
 void UTCPD_frs_mux_selection(int port, uint32_t cc1frssel, uint32_t cc2frssel)
 {
     int data; 
-	tcpc_addr_read16(port, NULL, UTCPD_MUXSEL, &data);
+	tcpc_addr_read16(port, (int) NULL, UTCPD_MUXSEL, &data);
 	data = (data & ~(CC2FRSS|CC1FRSS)) | ((cc1frssel<<25) | (cc2frssel<<29)); 
-    tcpc_addr_write32(port, NULL, UTCPD_MUXSEL, data);
+    tcpc_addr_write32(port, (int) NULL, UTCPD_MUXSEL, data);
 }
 
 /**
@@ -1128,7 +1128,7 @@ uint32_t UTCPD_Open(int port)
 
 	/* Enable PHY */
 	SYS->UTCPDCTL = SYS->UTCPDCTL | 0x02;
-	tcpc_addr_write16(port, NULL, UTCPD_PHYCTL, 0x03);	
+	tcpc_addr_write16(port, (int) NULL, UTCPD_PHYCTL, 0x03);
